requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
faiss-cpu
pypdf
tiktoken